% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_additional.R
\name{load_additional}
\alias{load_additional}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_additional(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "additional",
  add = F,
  selvar2 = c("patid", "enttype", "adid", "data1", "data2", "data3", "data4", "data5",
    "data6", "data7", "data8", "data9", "data10", "data11", "data12")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}

\item{bpp}{BiocParallal Multicore Parameters}
}
\description{
get the drugissue records and convert some fields to useful field types
}
