% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clinical.R
\name{load_clinical}
\alias{load_clinical}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_clinical(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "clinical",
  add = F,
  selvars2 = c("patid", "eventdate", "sysdate", "constype", "consid", "medcode", "sctid",
    "sctdescid", "sctexpression", "sctmaptype", "sctmapversion", "sctisindicative",
    "sctisassured", "staffid", "episode", "enttype", "adid")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}

\item{bpp}{BiocParallal Multicore Parameters}
}
\description{
get the drugissue records and convert some fields to useful field types
}
