% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_therapy.R
\name{load_therapy}
\alias{load_therapy}
\title{get the drugissue records and convert some fields to useful field types}
\usage{
load_therapy(
  pddir,
  dbf,
  ow = F,
  db = F,
  tab_name = "therapy",
  add = F,
  selvars2 = c("patid", "eventdate", "sysdate", "consid", "prodcode", "drugdmd",
    "staffid", "dosageid", "bnfcode", "qty", "numdays", "numpacks", "packtype",
    "issueseq", "prn")
)
}
\arguments{
\item{batch}{the name of the batch subdirectory}

\item{datadir}{the name of the data directory}

\item{bpp}{BiocParallal Multicore Parameters}
}
\description{
get the drugissue records and convert some fields to useful field types
}
