% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_control_sprint.R, R/make_controls_cohort.R
\name{make_control_cohort}
\alias{make_control_cohort}
\title{Make a control cohort using some inclusion and exclusion lists}
\usage{
make_control_cohort(dbn, exclude, include, db = T, silent = F)

make_control_cohort(dbn, exclude, include, db = T, silent = F)
}
\description{
Make a control cohort using some inclusion and exclusion lists

Make a control cohort using some inclusion and exclusion lists
}
