% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_control_sprint.R
\name{make_control_sprint}
\alias{make_control_sprint}
\title{Make a control cohort using some inclusion and exclusion lists}
\usage{
make_control_sprint(dbn, exclude, include, maxyear, db = T, silent = F)
}
\description{
Make a control cohort using some inclusion and exclusion lists
}
