% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_denomlinkage_db.R
\name{make_denom_db}
\alias{make_denom_db}
\title{make a demographic linkage database for control construction}
\usage{
make_denom_db(dbn, denomdir, linkdir, db = F)
}
\description{
make a demographic linkage database for control construction
}
